% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaSGM.R
\name{allKnownTFs}
\alias{allKnownTFs}
\title{specify the universe of recognized transcription factor gene symbols}
\usage{
allKnownTFs(source = "GO:DNAbindingTranscriptionFactorActivity")
}
\arguments{
\item{source}{A character string, currently only (and defaulted to) "GO:DNAbindingTranscriptionFactorActivity"}

\item{obj}{An object of class ModelBuilder, or any subclass}
}
\description{
specify the universe of recognized transcription factor gene symbols
}
