% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaSGM.R
\name{execute.footprint.strategy,trenaSGM-method}
\alias{execute.footprint.strategy,trenaSGM-method}
\alias{execute.footprint.strategy}
\title{build a model based on the specified footprint strategy}
\usage{
\S4method{execute.footprint.strategy}{trenaSGM}(obj, strategy)
}
\arguments{
\item{obj}{An object of class trenaSGM}

\item{footprint.strategy}{a list of parameters for footprint-driven model building}
}
\value{
A list with two data.frames: a model, and motif regions
}
\description{
build a model based on the specified footprint strategy
}
\examples{
if(interactive()){
  }

}
