% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoDatabaseModelBuilder.R
\docType{class}
\name{FimoDatabaseModelBuilder-class}
\alias{FimoDatabaseModelBuilder-class}
\alias{.FimoDatabaseModelBuilder}
\alias{FimoDatabaseModelBuilder}
\title{Create a FimoDatabaseModelBuilder object}
\usage{
FimoDatabaseModelBuilder(genomeName, targetGene, strategy, quiet = TRUE)
}
\arguments{
\item{genomeName}{hg38, mm10, ...}

\item{targetGene}{in same vocabulary as rownames in the expression matrix}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the FimoDatabaseModelBuilder class
}
\description{
tell us what we need to know
}
\examples{
if(interactive()){

  load(system.file(package="trenaSGM", "extdata", "mayo.tcx.RData"))
  fp.specs <- list(title="2000up.200down.fp",
                   type="database.fimos",
                   chrom="chr6",
                   tss=41163186,
                   upstream=2000,
                   downstream=200,
                   db.host="khaleesi.systemsbiology.net",
                   databases=list("brain_hint_16", "brain_hint_20", "brain_wellington_16", "brain_wellington_20"),
                   motifDiscovery="builtinFimo",
                   tfMapping=c("TFClass", "MotifDB"),
                   tfPrefilterCorrelation=0.2)
  fpc <- FimoDatabaseModelBuilder("hg38", "TREM2", fp.specs)
  }

}
