% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBuilder.R
\name{ModelBuilder}
\alias{ModelBuilder}
\title{Constructor for ModelBuilder}
\usage{
ModelBuilder(genomeName, targetGene, strategy, quiet = TRUE)
}
\arguments{
\item{genomeName}{hg38, mm10, ...}

\item{targetGene}{in same vocabulary as rownames in the expression matrix}

\item{strategy}{a list of options defining the model to be built}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the FootprintDatabaseModelBuilder class

An object of class ModelBuilder
}
\description{
Constructor for ModelBuilder
}
