% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{provide formatted console output as in the standard C library}
\usage{
printf(...)
}
\arguments{
\item{...}{format string and corresponding variables}
}
\value{
no return value, side effect is a formatted string printed to the R console
}
\description{
provide formatted console output as in the standard C library
}
