% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionsFimoModelBuilder.R
\docType{class}
\name{RegionsFimoModelBuilder-class}
\alias{RegionsFimoModelBuilder-class}
\alias{.RegionsFimoModelBuilder}
\alias{RegionsFimoModelBuilder}
\title{Create a RegionsFimoModelBuilder object}
\usage{
RegionsFimoModelBuilder(genomeName, targetGene, strategy, fimoClient,
  quiet = TRUE)
}
\arguments{
\item{genomeName}{hg38, mm10, ...}

\item{targetGene}{in same vocabulary as rownames in the expression matrix}

\item{fimoClient}{a live connection to a FimoServer}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the RegionsFimoModelBuilder class
}
\description{
tell us what we need to know
}
\examples{
if(interactive()){

  load(system.file(package="trenaSGM", "extdata", "mayo.tcx.RData"))
  spec <- list(title="2000up.200down.rmm",
               type="regions.fimo",
               tss=41163186,
               regions=data.frame(chrom="chr6", start=41162986, end=41165186, stringsAsFactors=FALSE)
               matrix=mtx,
               motifDiscovery="matchPWM",
               matchThreshold=90,
               tfMapping=list("MotifDB"),
               tfPrefilterCorrelation=0.2)
  rmmBuilder <- RegionsFimoModelBuilder("hg38", "TREM2", spec)
  build(rmmBuilder)
  }

}
