% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionsMotifMatchingModelBuilder.R
\docType{class}
\name{RegionsMotifMatchingModelBuilder-class}
\alias{RegionsMotifMatchingModelBuilder-class}
\alias{.RegionsMotifMatchingModelBuilder}
\alias{RegionsMotifMatchingModelBuilder}
\title{Create a RegionsMotifMatchingModelBuilder object}
\usage{
RegionsMotifMatchingModelBuilder(genomeName, targetGene, strategy,
  quiet = TRUE)
}
\arguments{
\item{genomeName}{hg38, mm10, ...}

\item{targetGene}{in same vocabulary as rownames in the expression matrix}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the RegionsMotifMatchingModelBuilder class
}
\description{
tell us what we need to know
}
\examples{
if(interactive()){

  load(system.file(package="trenaSGM", "extdata", "mayo.tcx.RData"))
  spec <- list(title="2000up.200down.rmm",
               type="regions.motifMatching",
               tss=41163186,
               regions=data.frame(chrom="chr6", start=41162986, end=41165186, stringsAsFactors=FALSE)
               matrix=mtx,
               motifDiscovery="matchPWM",
               matchThreshold=90,
               tfMapping=list("MotifDB"),
               tfPrefilterCorrelation=0.2)
  rmmBuilder <- RegionsMotifMatchingModelBuilder("hg38", "TREM2", spec)
  build(rmmBuilder)
  }

}
