% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundNumericColumns}
\alias{roundNumericColumns}
\title{truncate precision in data.frames}
\usage{
roundNumericColumns(tbl, digits, exponentialColumnNames = NA)
}
\arguments{
\item{tbl}{a data.frame}

\item{digits}{the number of places of decimal precision to preserve}

\item{exponentalColumnNames}{so that they can be preserved in exponential notation}
}
\value{
the transformed data.frame, with rows and columns intact
}
\description{
column by column rounding of data.frame numerical values, useful for readibility
}
