% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintDatabaseModelBuilder.R,
%   R/RegionsMotifMatchingModelBuilder.R, R/NoDnaModelBuilder.R,
%   R/RegionsFimoModelBuilder.R
\docType{methods}
\name{show,FootprintDatabaseModelBuilder-method}
\alias{show,FootprintDatabaseModelBuilder-method}
\alias{show}
\alias{show,RegionsMotifMatchingModelBuilder-method}
\alias{show}
\alias{show,NoDnaModelBuilder-method}
\alias{show}
\alias{show,RegionsFimoModelBuilder-method}
\alias{show}
\title{summarize the attributes specifying the creation of a trena gene regulatory model}
\usage{
\S4method{show}{FootprintDatabaseModelBuilder}(object)

\S4method{show}{RegionsMotifMatchingModelBuilder}(object)

\S4method{show}{NoDnaModelBuilder}(object)

\S4method{show}{RegionsFimoModelBuilder}(object)
}
\arguments{
\item{obj}{An object of class FootprintDatabaseModelBuilder}

\item{obj}{An object of class RegionsMotifMatchingModelBuilder}

\item{obj}{An object of class NoDnaModelBuilder}

\item{obj}{An object of class RegionsFimoModelBuilder}
}
\description{
summarize the attributes specifying the creation of a trena gene regulatory model

summarize the attributes specifying the creation of a trena gene regulatory model

summarize the attributes specifying the creation of a trena gene regulatory model

summarize the attributes specifying the creation of a trena gene regulatory model
}
