% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintDatabaseModelBuilder.R
\docType{methods}
\name{staged.build,FootprintDatabaseModelBuilder-method}
\alias{staged.build,FootprintDatabaseModelBuilder-method}
\alias{staged.build}
\title{create regulatory model of the gene, following all the specified options, one stage at a time, saving intermedate data}
\usage{
\S4method{staged.build}{FootprintDatabaseModelBuilder}(obj,
  stage = c("find.footprints", "associateTFs", "build.models"))
}
\arguments{
\item{obj}{An object of class FootprintDatabaseModelBuilder}

\item{stage}{a character string, one of "find.footprints", "associateTFs", "build.models"}

\item{strategy}{a list specifying all the options to build one or more models}
}
\value{
A list with two data.frames, "model" and "regulatoryRegions"
}
\description{
create regulatory model of the gene, following all the specified options, one stage at a time, saving intermedate data
}
\examples{
if(interactive()){
  fp.specs <- list(title="fp.2000up.200down",
                   type="database.footprints",
                   chrom="chr6",
                   tss=41163186,
                   start=(tss-2000),
                   downstream=(tss+200),
                   matrix
                   db.host="khaleesi.systemsbiology.net",
                   databases=list("brain_hint_20"),
                   motifDiscovery="builtinFimo",
                   tfMapping=c("TFClass", "MotifDB"),
                   tfPrefilterCorrelation=0.2)
  fpBuilder <- FootprintDatabaseModelBuilder("hg38", "TREM2", fp.specs, quiet=TRUE)
  load(system.file(package="trenaSGM", "extdata", "mayo.tcx.RData"))
  staged.build(fpBuilder, mtx, stage="find.footprints")
  }

}
