% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaSGM.R
\docType{class}
\name{trenaSGM-class}
\alias{trenaSGM-class}
\alias{.trenaSGM}
\alias{trenaSGM}
\title{Create a trenaSGM object}
\usage{
trenaSGM(genomeName, targetGene, quiet = TRUE)
}
\arguments{
\item{genomeName}{hg38, mm10, ...}

\item{targetGene}{in same vocabulary as rownames in the expression matrix}

\item{quiet}{do or do not print progress information}
}
\value{
An object of the trenaSGM class
}
\description{
tell us what we need to know
}
\examples{
if(interactive()){
  load(system.file(package="trenaSGM", "extdata", "mayo.tcx.RData"))
  pfms <- query(MotifDb, c("sapiens", "jaspar2018"))
  sgm <- trenaSGM("hg38", "TREM2", mtx, pfms,
                  strategies=list(footprints="5000.5000.remapped"),
                  tfCorrelation=0.2)
  }

}
