% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaSGM.R
\name{trimModel}
\alias{trimModel}
\title{trim an expansive model to a more credible one, using parameterized voting scheme}
\usage{
trimModel(tbl.model, tbl.reg, votesNeeded = 3, tf.keepers = c())
}
\arguments{
\item{tbl.model}{a data.frame, an expansive, unfiltered trena model data.frame, with perhaps > 200 tf rows}

\item{tbl.reg}{a data.frame,  included here only so that it can be returned, filtered to exclude binding sites for tfs
which did not make it into the trimmed model}

\item{votesNeeded, }{numeric, typically 2 or 3, at least this many votes, by any solver method, is required to keep}

\item{tf.keepers, }{a character vector,   keep these in the model no matter how well they performed.
a tf in the model}
}
\description{
trim an expansive model to a more credible one, using parameterized voting scheme
}
