% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaSGM.R
\name{allKnownTFs}
\alias{allKnownTFs}
\title{specify the universe of recognized transcription factor gene symbols}
\usage{
allKnownTFs(
  source = "GO:DNAbindingTranscriptionFactorActivity",
  identifierType = "geneSymbol"
)
}
\arguments{
\item{source}{A character string, currently only (and defaulted to) "GO:DNAbindingTranscriptionFactorActivity"}

\item{identifierType}{A string, one of "geneSymbol" (the default), "entrezGeneID", "ensemblGeneID"}

\item{obj}{An object of class ModelBuilder, or any subclass}
}
\description{
specify the universe of recognized transcription factor gene symbols
}
