% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintDatabaseModelBuilder.R,
%   R/RegionsMotifMatchingModelBuilder.R
\docType{methods}
\name{show,FootprintDatabaseModelBuilder-method}
\alias{show,FootprintDatabaseModelBuilder-method}
\alias{show}
\alias{show,RegionsMotifMatchingModelBuilder-method}
\alias{show}
\title{summarize the attributes specifying the creation of a trena gene regulatory model}
\usage{
\S4method{show}{FootprintDatabaseModelBuilder}(object)

\S4method{show}{RegionsMotifMatchingModelBuilder}(object)
}
\arguments{
\item{obj}{An object of class FootprintDatabaseModelBuilder}

\item{obj}{An object of class RegionsMotifMatchingModelBuilder}
}
\description{
summarize the attributes specifying the creation of a trena gene regulatory model

summarize the attributes specifying the creation of a trena gene regulatory model
}
