% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintDatabaseModelBuilder.R
\docType{methods}
\name{show,FootprintDatabaseModelBuilder-method}
\alias{show,FootprintDatabaseModelBuilder-method}
\alias{show}
\title{summarize the attributes specifying the creation of a trena gene regulatory model}
\usage{
\S4method{show}{FootprintDatabaseModelBuilder}(object)
}
\arguments{
\item{obj}{An object of class FootprintDatabaseModelBuilder}
}
\description{
summarize the attributes specifying the creation of a trena gene regulatory model
}
